pyproject.toml
[tool.poetry]
name = "tac-scenario-generator"
version = "0.1.0"
description = "Tool for generating tactical scenarios in a variety of videogames."
authors = ["fildred13 <fildred13@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tac_scenario_generator"}]

[tool.poetry.dependencies]
python = "^3.10"
PyAutoGUI = "^0.9.54"
easyocr = "^1.7.0"
Pillow = "^9.5.0"
thefuzz = {version = "^0.19.0", extras=["speedup"]}
torch = {version = "^2.0.1+cu118", source = "pytorch"}
torchvision = {version = "^0.15.2+cu118", source = "pytorch"}
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.poetry.group.codestyle]
optional = true

[tool.poetry.group.codestyle.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
